requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
langchain==0.3.7
langchain-core==0.3.15
langchain-openai==0.2.6
langgraph==0.2.39
langsmith>=0.1.14
supabase==2.9.1

# Phase 1: Authentication & Caching
redis==5.0.1
python-jose[cryptography]==3.3.0

# Health check dependency
requests==2.32.3

# Form data handling (required for FastAPI File/Form)
python-multipart==0.0.9

# Google Gemini API (for embeddings)
google-generativeai>=0.8.0

# Monitoring & Error Tracking (Sentry)
sentry-sdk[fastapi]==2.20.0

# Voice audio conversion (lighter than ffmpeg-python)
pydub>=0.25.1

# Phase 2: Hybrid Memory Recall (BM25 + reranking)
rank-bm25==0.2.2

# LLM Observability (LangFuse — self-hosted, free)
# Pin to v2.x — matches LangFuse server v2, avoids OTel conflicts with v4
langfuse>=2.50.0,<3.0.0

# Infrastructure Tracing (OpenTelemetry) — optional, disabled by default
# Pinned to avoid dependency resolution conflicts with langfuse v4
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-jaeger==1.21.0